Move SingleAccessMethodCompiler to io.trino.util
